changement int en size_t

+++ srcs/commands/operator/wallops.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   wallops.cpp                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: apommier <apommier@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sadjigui <sadjigui@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/16 17:09:38 by apommier          #+#    #+#             */
-/*   Updated: 2023/03/14 06:23:21 by apommier         ###   ########.fr       */
+/*   Updated: 2023/03/14 12:30:48 by sadjigui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ void	WALLOPS(std::string buffer, fdList &allFds, int userNbr)
 		return ;
 	}
 	msg = ":" + allFds.userData[userNbr].nickname + " WALLOPS :" + buffer.substr(splitBuff[1].size() + std::string::npos + 1) + "\r\n";
-	for (int i = 0; i < allFds.userData.size(); i++)
+	for (size_t i = 0; i < allFds.userData.size(); i++)
 	{
 		if (allFds.userData[i].op && allFds.userData[i].mode.s)
 			send(allFds.userData[i].fd, msg.c_str(), msg.size(), 0);
